+++ CHANGES.md
@@ -1,8 +1,9 @@
 # Changelog #
 
 * dev
-  * Improved checkers.
-  * Improved unit test coverage.
+  * Added `TypesettingQuotationMarks` check.
+  * Various checkers' code improvements.
+  * Added more unit tests.
   * Rephrased some report messages for clarity.
 
 * v1.2.0 (2021-07-21)